Updated upgrade script with the datatype and definition changes
Changed PostgreSQL-specific data types to data types defined in function/view specification in the following functions in the upgrade script:

sys.schema_name()
sys.has_perms_by_name()
sys.type_name()
sys.max_connections()

Also added missing columns to sys.computed_columns view and removed columns from sys.configurations view that were not part of the view definition in the upgrade script.

Task: BABELFISH-174
Author: Leszek Kurzyna <lkurzyna@amazon.com>

+++ contrib/babelfishpg_tsql/sql/sys_functions.sql
@@ -1784,7 +1784,7 @@ AS $$
 declare return_value integer;
 begin
     return_value := (select pg_trigger_depth());
-    RETURN return_value::integer;
+    RETURN return_value;
 EXCEPTION
     WHEN others THEN
         RETURN NULL;
